Refactor the RO
