+++ pkg/reconciler/rolloutorchestrator/controller.go
@@ -30,6 +30,7 @@ import (
 	spainformer "knative.dev/serving-progressive-rollout/pkg/client/injection/informers/serving/v1/stagepodautoscaler"
 	roreconciler "knative.dev/serving-progressive-rollout/pkg/client/injection/reconciler/serving/v1/rolloutorchestrator"
 	"knative.dev/serving-progressive-rollout/pkg/reconciler/common"
+	"knative.dev/serving-progressive-rollout/pkg/reconciler/rolloutorchestrator/rolloutmodes"
 	cfgmap "knative.dev/serving/pkg/apis/config"
 	revisioninformer "knative.dev/serving/pkg/client/injection/informers/serving/v1/revision"
 )
@@ -48,17 +49,20 @@ func NewController(
 	configStore := cfgmap.NewStore(logger.Named(common.ConfigStoreName))
 	configStore.WatchConfigs(cmw)
 
+	rolloutMode := rolloutmodes.NewRolloutModes(servingclient.Get(ctx), stagePodAutoscalerInformer.Lister())
 	c := &Reconciler{
 		client:                   servingclient.Get(ctx),
 		stagePodAutoscalerLister: stagePodAutoscalerInformer.Lister(),
 		deploymentLister:         deploymentInformer.Lister(),
 		revisionLister:           revisionInformer.Lister(),
+		rolloutMode:              rolloutMode,
 	}
+
 	opts := func(*controller.Impl) controller.Options {
 		return controller.Options{ConfigStore: configStore}
 	}
 
-	// Creat a controller.Impl that handles queuing and feeding work from
+	// Create a controller.Impl that handles queuing and feeding work from
 	// the queue through an implementation of controller.Reconciler for the RolloutOrchestrator.
 	impl := roreconciler.NewImpl(ctx, c, opts)
